urlparser: Thwart denial of service attacks
Thanks to freestyl3r for pointing these out!

Someone could DDOS us, either than directing us to some firewalled port
in which case, our long timeout left us unable to do anything.

The other case of attack would be to direct us to a stream url, in
which case, the retrieval would never end.

For both of these cases, we added a HEAD request with a small timeout.
Right now it seems to not need anything else, as both attacks
throw an exception.

If needed later, we can use HEAD.headers to filter anything else out.