Always guess encoding in urlparser
There are some pesky sites that don't report the right encoding in
their headers, and instead rely on the http-equiv meta to set it.
The fact that browsers are smart enough to guess don't make you
right to ommit headers people.. sigh!

Instead of parsing the whole html, I'm using charade that comes with
requests to always guess the header. Great results so far.

Also, if the content is not marked as (x)html or xml, just
output the content type and stop there.