Fix color escape in url for urlparse. GitHub bots.
There were times when a color coded line containing a url resulted
in a url with illegal characters. Fixed by stripping colors.

We are using GitHub bots to report pushes on the channel, so
we need to ignore their join/parts. urlparser does not ignore them
because we find its finding the url useful.