Make bot retry to take back his nick
Sometimes on a reconnection the old bot will be left behind
hogging the nickname. Now it will try to regain its nick
after two minutes.

Also silenced outgoing NICK messages from the bot in moolog
because it was resulting in duplicate entries.
Now we are just logging the server answer to nick changes.