urlparser: Let HEAD raise an exception. Fixes #19
The initial HEAD request that urlparser makes is there so it can
find problems before we make an actual request.
Before now it only looked into the content-type. But we saw
that it's possible to get an error response with a valid content-type.

Now we are letting the request raise an error if it results in anything
but an ok or a redirect.

As for #19, since mpd streams return a 400 error on a HEAD, this should
suffice.